RWA-5084 test fix

+++ src/testUtils/java/uk/gov/hmcts/reform/wataskmanagementapi/services/AuthorizationProvider.java
@@ -48,13 +48,21 @@ public class AuthorizationProvider {
     private IdamServiceApi idamServiceApi;
     @Autowired
     private AuthTokenGenerator serviceAuthTokenGenerator;
+    @Value("${idam.test.userCleanupEnabled:true}")
+    private boolean testUserDeletionEnabled;
 
     public void deleteAccount(String username) {
-        try {
-            log.info("Deleting test account '{}'", username);
-            idamServiceApi.deleteTestUser(username);
-        } catch (FeignException e) {
-            log.error("Failed to delete test account '{}'", username, e);
+
+        if (testUserDeletionEnabled) {
+            //If error is thrown while deleting the user, it will be caught and logged
+            try {
+                log.info("Deleting test account '{}'", username);
+                idamServiceApi.deleteTestUser(username);
+            } catch (FeignException e) {
+                log.error("Failed to delete test account '{}'", username, e);
+            }
+        } else {
+            log.info("Test User deletion feature flag was not enabled, user '{}' was not deleted", username);
         }
     }
